refactor: single-home staged path ownership
Route stage_to_ai_path and folder sanitization through lib/processing_paths so the writer, recovery seam, and legacy helper cannot drift. Also make find_orphaned_downloads take one explicit local-path contract and keep blocked post-move detail formatting out of lib/quality.py.
