+++ lib/quality.py
@@ -13,8 +13,6 @@ from typing import Any, Literal, Optional
 
 import msgspec
 
-from lib.download_recovery import classify_processing_path
-
 QUALITY_UPGRADE_TIERS = "lossless,mp3 v0,mp3 320"
 QUALITY_LOSSLESS = "lossless"
 
@@ -3057,22 +3055,17 @@ def find_orphaned_downloads(
     db_rows: list[dict[str, Any]],
     active_transfers: set[tuple[str, str]],
     *,
-    existing_local_paths: set[str] | None = None,
-    staging_dir: str | None = None,
-    slskd_download_dir: str | None = None,
+    existing_local_paths: set[str] | None,
 ) -> list[OrphanInfo]:
     """Detect downloading rows whose slskd transfers no longer exist. Pure — no I/O.
 
     Args:
         db_rows: album_requests rows (must include status, active_download_state).
         active_transfers: set of (username, filename) tuples from slskd API.
-        existing_local_paths: optional set of persisted ``current_path`` values
-            that still exist on disk, supplied by the caller when local
-            filesystem visibility is available.
-        staging_dir: optional staging root for classifying persisted
-            ``current_path`` values in repair output.
-        slskd_download_dir: optional download root for classifying persisted
-            ``current_path`` values in repair output.
+        existing_local_paths: set of persisted ``current_path`` values that
+            still exist on disk, supplied by the caller when local filesystem
+            visibility is available. Pass ``None`` when the caller cannot
+            inspect local processing paths.
 
     Returns OrphanInfo for each downloading row where NONE of its files
     appear in active_transfers, plus ``blocked_post_move`` when a row is
@@ -3103,22 +3096,12 @@ def find_orphaned_downloads(
                 and existing_local_paths is not None
                 and current_path not in existing_local_paths
             ):
-                path_detail = "processing path"
-                if staging_dir and slskd_download_dir:
-                    path_detail = classify_processing_path(
-                        current_path=current_path,
-                        artist=str(row.get("artist_name") or ""),
-                        title=str(row.get("album_title") or ""),
-                        year=str(row.get("year") or ""),
-                        request_id=int(row["id"]),
-                        staging_dir=staging_dir,
-                        slskd_download_dir=slskd_download_dir,
-                    ).display_name
                 issues.append(OrphanInfo(
                     request_id=row["id"],
                     issue_type="blocked_post_move",
                     detail=(
-                        f"persisted {path_detail} missing after local processing: "
+                        "persisted processing path missing after local "
+                        "processing: "
                         f"{current_path}"
                     ),
                 ))
